cargo clippy

+++ parser/src/parser/ast.rs
@@ -26,8 +26,8 @@ trait GetNode {
 impl From<Node> for SourceSpan {
     fn from(val: Node) -> Self {
         Self::new(
-            SourceOffset::from(val.start as usize),
-            SourceOffset::from(val.len() as usize),
+            SourceOffset::from(val.start),
+            SourceOffset::from(val.len()),
         )
     }
 }
